+++ R/power.t.test.R
@@ -13,7 +13,7 @@
 #' @param type Type of t test
 #' @param alternative One- or two-sided test
 #' @param df.method Method for calculating the degrees of default. Possibilities are welch (the default) or classical.
-#' @param strict Use strict interpretation in two-sided case
+#' @param strict Use strict interpretation in two-sided case. Defaults to TRUE unlike the standard power.t.test function.
 #'
 #' @return  Object of class \code{power.htest}, a list of the arguments (including the computed one)
 #' augmented with \code{method} and \code{note} elements.
@@ -22,7 +22,7 @@
 #' and that parameter is determined from the others. Notice that the last two have non-NULL defaults
 #' so NULL must be explicitly passed if you want to compute them.
 #'
-#' If \code{strict = TRUE} is used, the power will include the probability
+#' The default \code{strict = TRUE} ensures that the power will include the probability
 #' of rejection in the opposite direction of the true effect, in the
 #' two-sided case. Without this the power will be half the
 #' significance level if the true difference is zero.
@@ -50,7 +50,7 @@ power_t_test <-
             type = c("two.sample", "one.sample", "paired"),
             alternative = c("two.sided", "one.sided"),
             df.method = c("welch", "classical"),
-            strict = FALSE)
+            strict = TRUE)
 {
   type <- match.arg(type)
   if (type == "two.sample") {