+++ NEWS.md
@@ -1,5 +1,9 @@
 # Next version
 
+*   Changed the internals of gkgamma to reduce the risk of integer overflow (thanks to Julián Darío Otero Niño)
+
+# MESS 0.5.7
+
 *   Added the plr function
 *   Added usd function for unbiased sd estimation
 *   Fixed a bug in the power_mcnemar_test function